# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/alunos": {
            "get": {
                "description": "Retorna uma lista de todos os alunos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alunos"
                ],
                "summary": "Retorna todos os alunos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Aluno"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Adiciona um novo aluno ao banco de dados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alunos"
                ],
                "summary": "Cadastra um novo aluno",
                "parameters": [
                    {
                        "description": "Aluno",
                        "name": "aluno",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Aluno"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Aluno"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.Resposta"
                        }
                    }
                }
            }
        },
        "/alunos/cpf/{cpf}": {
            "get": {
                "description": "Retorna os dados de um aluno específico pelo CPF",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alunos"
                ],
                "summary": "Retorna um aluno pelo CPF",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Aluno CPF",
                        "name": "cpf",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Aluno"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controllers.Resposta"
                        }
                    }
                }
            }
        },
        "/alunos/{id}": {
            "get": {
                "description": "Retorna os dados de um aluno específico pelo ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alunos"
                ],
                "summary": "Retorna um aluno pelo ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Aluno ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Aluno"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controllers.Resposta"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove um aluno do banco de dados pelo ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alunos"
                ],
                "summary": "Deleta um aluno pelo ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Aluno ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.Resposta"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controllers.Resposta"
                        }
                    }
                }
            },
            "patch": {
                "description": "Atualiza os dados de um aluno específico pelo ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alunos"
                ],
                "summary": "Edita um aluno pelo ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Aluno ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Aluno",
                        "name": "aluno",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Aluno"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Aluno"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.Resposta"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controllers.Resposta"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.Resposta": {
            "type": "object",
            "properties": {
                "mensagem": {
                    "type": "string"
                }
            }
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "models.Aluno": {
            "type": "object",
            "properties": {
                "cpf": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string"
                },
                "rg": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        }
    }
}